# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ticket-assignments": {
            "get": {
                "description": "Get a list of all ticket assignments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-assignments"
                ],
                "summary": "Get all ticket assignments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/requests.TicketAssignmentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-assignments"
                ],
                "summary": "Create a new ticket assignment",
                "parameters": [
                    {
                        "description": "Ticket Assignment Data",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateTicketAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketAssignmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ticket-assignments/{id}": {
            "get": {
                "description": "Get a ticket assignment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-assignments"
                ],
                "summary": "Get a ticket assignment by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketAssignmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ticket assignment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-assignments"
                ],
                "summary": "Update a ticket assignment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Assignment Data",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateTicketAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketAssignmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket assignment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-assignments"
                ],
                "summary": "Delete a ticket assignment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/tickets": {
            "get": {
                "description": "Get a list of all tickets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get all tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/requests.TicketResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Create a new ticket",
                "parameters": [
                    {
                        "description": "Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Get a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get a ticket by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ticket by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Update a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Delete a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "validation": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "requests.CreateTicketAssignmentRequest": {
            "type": "object",
            "properties": {
                "id_admin": {
                    "type": "integer",
                    "example": 2
                },
                "id_ticket": {
                    "type": "integer",
                    "example": 1
                },
                "tanggal_ditugaskan": {
                    "type": "string",
                    "example": "2025-11-07T10:00:00Z"
                }
            }
        },
...
        "requests.TicketAssignmentResponse": {
            "description": "TicketAssignmentResponse represents a ticket assignment with related ticket and admin info (no null fields)",
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/requests.UserSimpleResponse"
                },
                "id_admin": {
                    "type": "integer"
                },
                "id_assignment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "tanggal_ditugaskan": {
                    "type": "string"
                },
                "ticket": {
                    "$ref": "#/definitions/requests.TicketResponse"
                }
            }
        },
...
        "requests.TicketCreateRequest": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string",
                    "example": "Saya tidak bisa login ke akun saya."
                },
                "id_category": {
                    "type": "integer",
                    "example": 1
                },
                "id_priority": {
                    "type": "integer",
                    "example": 1
                },
                "id_status": {
                    "type": "integer",
                    "example": 1
                },
                "id_user": {
                    "type": "integer",
                    "example": 1
                },
                "judul": {
                    "type": "string",
                    "example": "Tidak bisa login"
                },
                "kode_tiket": {
                    "type": "string",
                    "example": "TCKT-001"
                },
                "tipe_pengaduan": {
                    "description": "Allowed: pelanggan, penjual",
                    "type": "string",
                    "enum": [
                        "pelanggan",
                        "penjual"
                    ],
                    "example": "pelanggan"
                }
            }
        },
        "requests.TicketResponse": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "id_category": {
                    "type": "integer"
                },
                "id_priority": {
                    "type": "integer"
                },
                "id_status": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "kode_tiket": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "tanggal_diperbarui": {
                    "type": "string"
                },
                "tipe_pengaduan": {
                    "type": "string"
                }
            }
        },
...
        "requests.UserSimpleResponse": {
            "description": "UserSimpleResponse represents a minimal user info for assignment admin",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id_user": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}